% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_p_value.R
\name{scan_p_value}
\alias{scan_p_value}
\title{Title: Calculate Scan P-Value}
\usage{
scan_p_value(A, num_permutations = 1000)
}
\arguments{
\item{A}{A numeric matrix where each row represents a cell pair. The first column represents the x-axis coordinates of the midpoint of this cell pair, the second column represents the y-axis coordinates, and the third column represents the communication status (1 for communication, 0 for no communication).}

\item{num_permutations}{Number of permutations to estimate the null distribution. Default is 1000.}
}
\value{
A p-value for the observed communication pattern.
}
\description{
Description: This function calculates the p-value by performing a scan based on a matrix of communication statuses between cell pairs.
}
\examples{
A <- matrix(c(1, 1, 1,
              2, 2, 0,
              3, 3, 1,
              4, 4, 0,
              5, 5, 0,
              6, 6, 0), 
            ncol = 3, byrow = TRUE)
scan_p_value(A, num_permutations = 1000)
}
